refactor: Rule은 사다리를 생성 규칙을 정하므로 패키지를 따로 rule로 분리
