+++ src/main/java/laddergame/controller/rule/RandomCreateRule.java
@@ -1,4 +1,4 @@
-package laddergame.domain.rule;
+package laddergame.controller.rule;
 
 import laddergame.util.RandomGenerator;
