+++ src/main/java/laddergame/controller/rule/Rule.java
@@ -1,4 +1,4 @@
-package laddergame.domain.rule;
+package laddergame.controller.rule;
 
 public interface Rule {
     boolean canCreate();
