+++ src/main/java/laddergame/domain/LineGenerator.java
@@ -1,7 +1,6 @@
 package laddergame.domain;
 
-import laddergame.domain.Line;
-import laddergame.domain.rule.Rule;
+import laddergame.controller.rule.Rule;
 
 public class LineGenerator {
     public static Line lineGenerate(int numberOfPerson, Rule rule) {
