+++ src/main/java/laddergame/util/OutputView.java
@@ -1,5 +1,6 @@
 package laddergame.util;
 
+import laddergame.controller.LadderGameResult;
 import laddergame.domain.*;
 
 import java.util.ArrayList;
